stream_processor/requirements.txt
streamlink
ffmpeg-python
redis
